Decouple ABI version from API version
Set the ABI version (used in so/DLL name) to 1 and increment it on an ABI
break (but only once within one release cycle). This way we don't need to
increment it on every patch release but also don't need to make a new major
release on every ABI break.

Keep the API version as is (using semantic versioning).