Avoid too big icons when rendering via the Quick image provider
This is basically applying/using the change from bb2030c9 for the Quick
image provider as well because it seems that the requested size already
takes scaling into account (so icons would be too big if the Quick image
provider would scale in addition).

When no size is requested (e.g. the case for the Plasmoid) then only an
initially assigned default size is used and scaling will still be applied
as before. This change however lowers the default for the default size to
32 because 64 is probably bigger than what is needed in most cases
(considering it will still be scaled as necassary).